+++ app/src/main/java/com/bridger/ui/scanner/DeviceListAdapter.java
@@ -1,14 +1,21 @@
 package com.bridger.ui.scanner;
 
+import android.content.Context;
+import android.content.Intent;
 import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.ListAdapter;
 
+import com.bridger.ui.device.DeviceActivity;
+
 public class DeviceListAdapter extends ListAdapter<DeviceScanResult, DeviceListViewHolder> {
 
-    public DeviceListAdapter() {
+    private final Context context;
+
+    public DeviceListAdapter(Context context) {
         super(DIFF_CALLBACK);
+        this.context = context;
     }
 
     @NonNull
@@ -21,6 +28,12 @@ public class DeviceListAdapter extends ListAdapter<DeviceScanResult, DeviceListV
     public void onBindViewHolder(@NonNull DeviceListViewHolder holder, int position) {
         DeviceScanResult device = getItem(position);
         holder.bind(device);
+        holder.itemView.setOnClickListener(v -> {
+            Intent intent = new Intent(context, DeviceActivity.class);
+            intent.putExtra(DeviceActivity.EXTRA_DEVICE_ADDRESS, device.getAddress());
+            intent.putExtra(DeviceActivity.EXTRA_DEVICE_NAME, device.getName()); // Pass the device name
+            context.startActivity(intent);
+        });
     }
 
     private static final DiffUtil.ItemCallback<DeviceScanResult> DIFF_CALLBACK =